+++ trikot-foundation/trikotFoundation/src/commonMain/kotlin/com/mirego/trikot/foundation/concurrent/dispatchQueue/TrikotDispatchQueues.kt
@@ -1,6 +1,6 @@
 package com.mirego.trikot.foundation.concurrent.dispatchQueue
 
-expect class OperationDispatchQueue() : TrikotDispatchQueue
-
-@Deprecated("Streams subscription concurrency is now handled by a serial queue in PublishSubject")
-expect class SerialSubscriptionDispatchQueue() : TrikotDispatchQueue
+expect class OperationDispatchQueue() : TrikotDispatchQueue {
+    override fun isSerial(): Boolean
+    override fun dispatch(block: DispatchBlock)
+}