+++ trikot-foundation/trikotFoundation/src/commonMain/kotlin/com/mirego/trikot/foundation/concurrent/dispatchQueue/UIThreadDispatchQueue.kt
@@ -1,3 +1,6 @@
 package com.mirego.trikot.foundation.concurrent.dispatchQueue
 
-expect class UIThreadDispatchQueue() : TrikotDispatchQueue
+expect class UIThreadDispatchQueue() : TrikotDispatchQueue {
+    override fun isSerial(): Boolean
+    override fun dispatch(block: DispatchBlock)
+}