+++ trikot-foundation/trikotFoundation/src/nativeMain/kotlin/com/mirego/trikot/foundation/concurrent/dispatchQueue/UIThreadDispatchQueue.kt
@@ -3,14 +3,14 @@ package com.mirego.trikot.foundation.concurrent.dispatchQueue
 import platform.Foundation.NSThread
 import platform.darwin.dispatch_async
 import platform.darwin.dispatch_get_main_queue
-import kotlin.native.concurrent.AtomicInt
+import kotlin.concurrent.AtomicInt
 
 actual class UIThreadDispatchQueue actual constructor() : TrikotDispatchQueue {
     private val count = AtomicInt(0)
 
-    override fun isSerial() = true
+    actual override fun isSerial() = true
 
-    override fun dispatch(block: DispatchBlock) {
+    actual override fun dispatch(block: DispatchBlock) {
         val currentCount = count.addAndGet(1)
         if (currentCount == 1 && NSThread.isMainThread) {
             runQueueTask(block)