+++ trikot-viewmodels-declarative/viewmodels-declarative/src/commonMain/kotlin/com/mirego/trikot/viewmodels/declarative/controller/VMDPlatformViewModelController.kt
@@ -1,5 +1,5 @@
 package com.mirego.trikot.viewmodels.declarative.controller
 
 expect abstract class VMDPlatformViewModelController() {
-    open fun onCleared()
+    protected open fun onCleared()
 }