+++ trikot-viewmodels/viewmodels/src/commonMain/kotlin/com/mirego/trikot/viewmodels/lifecycle/ApplicationStatePublisher.kt
@@ -1,5 +1,8 @@
 package com.mirego.trikot.viewmodels.lifecycle
 
 import org.reactivestreams.Publisher
+import org.reactivestreams.Subscriber
 
-expect class ApplicationStatePublisher() : Publisher<ApplicationState>
+expect class ApplicationStatePublisher() : Publisher<ApplicationState> {
+    override fun subscribe(s: Subscriber<in ApplicationState>)
+}