+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Upcoming
 
+### Updates
+
+- Kingfisher to `7.10.1`
+
 ### Breaking Changes
 
 - [VMD] Change parameters order for some Compose components