+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Upcoming
 
+
 ### Updates
 
 - Kingfisher to `7.10.1`
@@ -13,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Breaking Changes
 
 - [VMD] Change parameters order for some Compose components
+- [VMD] Change VMDTextField's signature on Android
 
 ### Updates