+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Breaking Changes
 
 - [VMD] Change parameters order for some Compose components
+- [VMD] Change VMDTextField's signature on Android
 
 ### Updates